This commit was manufactured by cvs2git to create tag 'JetMETCorrections- METPUSubtraction-MM_2013Jun07'.
Sprout from master 2013-02-25 15:08:39 UTC Christian Veelken <sha1-5c72da6f595cce9b6b48aff6d56f01e9beb4aad1@cern.ch> '- added protection in case "lepton" is close to jet with jetPt >> leptonPt:'
Cherrypick from master 2013-06-07 14:26:14 UTC Matthieu Pierre Marionneau <sha1-9ed410376145397a6c346fc09a7f2c12b408240e@cern.ch> 'changed format of association maps':
    interface/noPileUpMEtAuxFunctions.h
    src/noPileUpMEtAuxFunctions.cc

+++ src/noPileUpMEtAuxFunctions.cc
@@ -17,7 +17,7 @@ int isVertexAssociated(const reco::PFCandidate& pfCandidate,
     vtxAssociationType = 1;
     for ( PFCandidateToVertexAssociationMap::const_iterator pfCandToVertexAssociation = pfCandToVertexAssociations.begin();
           pfCandToVertexAssociation != pfCandToVertexAssociations.end(); ++pfCandToVertexAssociation ) {
-      typedef std::vector<std::pair<reco::PFCandidateRef, float> > PFCandidateQualityPairVector;
+      typedef std::vector<std::pair<reco::PFCandidateRef, int> > PFCandidateQualityPairVector;
       const PFCandidateQualityPairVector& pfCandidates_vertex = pfCandToVertexAssociation->val;
       for ( PFCandidateQualityPairVector::const_iterator pfCandidate_vertex = pfCandidates_vertex.begin();
             pfCandidate_vertex != pfCandidates_vertex.end(); ++pfCandidate_vertex ) {
@@ -44,7 +44,7 @@ reversedPFCandidateToVertexAssociationMap reversePFCandToVertexAssociation(const
   for ( PFCandidateToVertexAssociationMap::const_iterator pfCandToVertexAssociation = pfCandToVertexAssociations.begin();
 	pfCandToVertexAssociation != pfCandToVertexAssociations.end(); ++pfCandToVertexAssociation ) {
     const reco::VertexRef& vertex = pfCandToVertexAssociation->key;
-    typedef std::vector<std::pair<reco::PFCandidateRef, float> > PFCandidateQualityPairVector;
+    typedef std::vector<std::pair<reco::PFCandidateRef, int> > PFCandidateQualityPairVector;
     const PFCandidateQualityPairVector& pfCandidates_vertex = pfCandToVertexAssociation->val;
     for ( PFCandidateQualityPairVector::const_iterator pfCandidate_vertex = pfCandidates_vertex.begin();
 	  pfCandidate_vertex != pfCandidates_vertex.end(); ++pfCandidate_vertex ) {
